Доработал, исправил, и добавил комментарии для функций
